feat: set ts strict mode on

+++ pages/api/post/[id].ts
@@ -1,8 +1,15 @@
+import { NextApiRequest, NextApiResponse } from "next";
 import prisma from "../../../lib/prisma";
 
 // DELETE /api/post/:id
-export default async function handle(req, res) {
+export default async function handle(
+  req: NextApiRequest,
+  res: NextApiResponse
+) {
   const postId = req.query.id;
+  if (typeof postId !== "string") {
+    return res.status(400).json({ message: "`id` must be a string" });
+  }
   if (req.method === "DELETE") {
     const post = await prisma.post.delete({
       where: { id: postId },
